# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Freeable Domains API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Trigger a new deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Deployment triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get a deployment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/rollback:
    post:
      operationId: rollbackDeployment
      tags: [deployments]
      summary: Rollback to this deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rollback successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/logs:
    get:
      operationId: getDeploymentLogs
      tags: [builds]
      summary: Get build/deployment logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"

  # ── Domains ───────────────────────────────────────────────────────────────
  /domains/search:
    get:
      operationId: searchDomains
      tags: [domains]
      summary: Search & check domain availability
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DomainSearchResult"
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: List all registered domains
      responses:
        "200":
          description: Domain list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
    post:
      operationId: registerDomain
      tags: [domains]
      summary: Register a domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
  /domains/{id}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Get domain details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Domain details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Delete / release a domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /domains/{id}/attach:
    post:
      operationId: attachDomainToProject
      tags: [domains]
      summary: Attach domain to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainAttach"
      responses:
        "200":
          description: Attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"

  # ── DNS Records ───────────────────────────────────────────────────────────
  /domains/{id}/dns:
    get:
      operationId: listDnsRecords
      tags: [dns]
      summary: List DNS records for a domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DNS records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DnsRecord"
    post:
      operationId: createDnsRecord
      tags: [dns]
      summary: Create a DNS record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DnsRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsRecord"
  /dns/{id}:
    put:
      operationId: updateDnsRecord
      tags: [dns]
      summary: Update a DNS record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DnsRecordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsRecord"
    delete:
      operationId: deleteDnsRecord
      tags: [dns]
      summary: Delete a DNS record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Environment Variables ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Deployment:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, cloning, installing, building, deploying, ready, failed, cancelled, rolled_back]
        url:
          type: ["string", "null"]
        commitSha:
          type: ["string", "null"]
        commitMessage:
          type: ["string", "null"]
        branch:
          type: string
        buildDurationSeconds:
          type: ["number", "null"]
        triggeredBy:
          type: string
          enum: [manual, github_push, api, rollback]
        isProduction:
          type: boolean
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - projectId
        - status
        - branch
        - triggeredBy
        - isProduction
        - createdAt
    DeploymentInput:
      type: object
      properties:
        branch:
          type: string
        commitSha:
          type: string
        isProduction:
          type: boolean
      required: []

    # ── Logs ────────────────────────────────────────────────────────────────
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        deploymentId:
          type: integer
        level:
          type: string
          enum: [info, warn, error, success, debug]
        stage:
          type: string
          enum: [clone, install, build, deploy, verify]
        message:
          type: string
        timestamp:
          type: string
      required: [id, deploymentId, level, stage, message, timestamp]

    # ── Domains ─────────────────────────────────────────────────────────────
    Domain:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tld:
          type: string
        fullDomain:
          type: string
        status:
          type: string
          enum: [active, pending, expired, suspended, transferring]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        registeredAt:
          type: string
        expiresAt:
          type: string
        autoRenew:
          type: boolean
        sslEnabled:
          type: boolean
        sessionId:
          type: ["string", "null"]
      required:
        - id
        - name
        - tld
        - fullDomain
        - status
        - registeredAt
        - expiresAt
        - autoRenew
        - sslEnabled
    DomainInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        tld:
          type: string
          minLength: 1
        projectId:
          type: integer
        autoRenew:
          type: boolean
        sessionId:
          type: string
      required: [name, tld]
    DomainSearchResult:
      type: object
      properties:
        name:
          type: string
        tld:
          type: string
        fullDomain:
          type: string
        available:
          type: boolean
        premium:
          type: boolean
      required: [name, tld, fullDomain, available, premium]
    DomainAttach:
      type: object
      properties:
        projectId:
          type: integer
      required: [projectId]

    # ── DNS Records ─────────────────────────────────────────────────────────
    DnsRecord:
      type: object
      properties:
        id:
          type: integer
        domainId:
          type: integer
        type:
          type: string
          enum: [A, AAAA, CNAME, TXT, MX, NS, SRV, CAA, PTR, ALIAS]
        name:
          type: string
        value:
          type: string
        ttl:
          type: integer
        priority:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, domainId, type, name, value, ttl, createdAt]
    DnsRecordInput:
      type: object
      properties:
        type:
          type: string
          enum: [A, AAAA, CNAME, TXT, MX, NS, SRV, CAA, PTR, ALIAS]
        name:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
        ttl:
          type: integer
        priority:
          type: integer
      required: [type, name, value]
    DnsRecordUpdate:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        value:
          type: string
        ttl:
          type: integer
        priority:
          type: integer

    # ── Environment Variables ───────────────────────────────────────────────
...