# lib/api-zod/src/generated/types/deploymentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Freeable Domains API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeploymentInput {
  branch?: string;
  commitSha?: string;
  isProduction?: boolean;
}