# lib/api-zod/src/generated/types/dnsRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Freeable Domains API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DnsRecordType } from './dnsRecordType';

export interface DnsRecord {
  id: number;
  domainId: number;
  type: DnsRecordType;
  name: string;
  value: string;
  ttl: number;
  /** @nullable */
  priority?: number | null;
  createdAt: string;
}