# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Freeable Domains API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DomainStatus } from './domainStatus';

export interface Domain {
  id: number;
  name: string;
  tld: string;
  fullDomain: string;
  status: DomainStatus;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  registeredAt: string;
  expiresAt: string;
  autoRenew: boolean;
  sslEnabled: boolean;
  /** @nullable */
  sessionId?: string | null;
}