# lib/api-zod/src/generated/types/domainInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Freeable Domains API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DomainInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  tld: string;
  projectId?: number;
  autoRenew?: boolean;
  sessionId?: string;
}